+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Features:
 
 Bug Fixes:
 
+- Enable Ctrl+S (and File > Save) keyboard shortcut to work in the CMake Cache Editor UI. [#4000](https://github.com/microsoft/vscode-cmake-tools/issues/4000)
 - Fix "CMake: Run Without Debugging" not changing the working directory when the build directory changes. [#4549](https://github.com/microsoft/vscode-cmake-tools/issues/4549)
 - Fix CMake script path links not working in CHS/CSY/FRA/PLK locales due to localized quotes. [#4383](https://github.com/microsoft/vscode-cmake-tools/issues/4383)
 - Fix user-level tasks defined in `~/.config/Code/User/tasks.json` causing infinite spinner. [#4659](https://github.com/microsoft/vscode-cmake-tools/pull/4659)